Organize project folder and prepare template
