+++ Makefile
@@ -1,2 +1,2 @@
 start:
-	go run main.go
+	go run cmd/main.go
