+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/psinthorn/go_smallsite/app"
+
+func main() {
+	app.Start()
+}
