+++ templates/base.layout.html
@@ -1,0 +1,24 @@
+{{define "base"}}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
+    <title>Small Site: Home Page</title>
+    {{block "css" .}}
+
+    {{end}}
+</head>
+<body>
+    {{block "content" .}}
+
+    {{end}}
+
+
+    {{block "js" .}}
+
+    {{end}}
+</body>
+</html>
+{{end}}
