+++ templates/home.page.html
@@ -1,11 +1,11 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Small Site: Home Page</title>
-</head>
-<body>
-        <h1>Home Page</h1>
-</body>
-</html>
+{{template "base" .}}
+
+        {{define "contebt"}}
+                <div class="container">
+                        <div class="ro">
+                                <div class="col">
+                                        <h1>Home Page</h1>
+                                </div>
+                        </div>
+                </div>
+        {{end}}
